+++ pkg/provider/provider_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	"github.com/cyberark/conjur-k8s-csi-provider/pkg/conjur"
+	"github.com/cyberark/conjur-k8s-csi-provider/pkg/k8s"
 	"github.com/stretchr/testify/assert"
 	"sigs.k8s.io/secrets-store-csi-driver/provider/v1alpha1"
 )
@@ -23,10 +24,11 @@ func (c *mockConjurClient) GetSecrets(jwt string, secretIds []string) (map[strin
 
 func TestMount(t *testing.T) {
 	testCases := []struct {
-		description   string
-		req           *v1alpha1.MountRequest
-		conjurFactory conjur.ClientFactory
-		assertions    func(*testing.T, *v1alpha1.MountResponse, error)
+		description        string
+		req                *v1alpha1.MountRequest
+		conjurFactory      conjur.ClientFactory
+		getAnnotationsFunc k8s.GetPodAnnotationsFunc
+		assertions         func(*testing.T, *v1alpha1.MountResponse, error)
 	}{
 		{
 			description: "throws error decoding invalid attributes",
@@ -41,7 +43,7 @@ func TestMount(t *testing.T) {
 		{
 			description: "throws error decoding invalid attributes",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"conjur.org/configurationVersion": "0.2.0"}`,
+				Attributes: `{"conjur.org/configurationVersion": "0.3.0"}`,
 			},
 			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
 				assert.Nil(t, resp)
@@ -79,9 +81,9 @@ func TestMount(t *testing.T) {
 			},
 		},
 		{
-			description: "throws error when secrets attribute not included or empty",
+			description: "throws error when secrets attribute not included or empty (v0.1.0)",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"secrets":"","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Attributes: `{"conjur.org/configurationVersion":"0.1.0","secrets":"","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
 			},
 			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
 				assert.Nil(t, resp)
@@ -89,9 +91,41 @@ func TestMount(t *testing.T) {
 			},
 		},
 		{
-			description: "throws error when secrets attribute improperly formatted",
+			description: "throws error when secrets attribute improperly formatted (v0.1.0)",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"secrets":"invalid","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Attributes: `{"conjur.org/configurationVersion":"0.1.0","secrets":"invalid","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+			},
+			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
+				assert.Nil(t, resp)
+				assert.Contains(t, err.Error(), "Failed to unmarshal secrets spec")
+			},
+		},
+		{
+			description: "throws error when secrets attribute not included or empty (v0.2.0)",
+			req: &v1alpha1.MountRequest{
+				Attributes: `{"sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+			},
+			getAnnotationsFunc: func(namespace string, podName string) (map[string]string, error) {
+				return map[string]string{
+					"conjur.org/secrets":    "",
+					"some-other-annotation": "some-value",
+				}, nil
+			},
+			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
+				assert.Nil(t, resp)
+				assert.Contains(t, err.Error(), "CKCP034 Annotation \"conjur.org/secrets\" missing or empty")
+			},
+		},
+		{
+			description: "throws error when secrets attribute improperly formatted (v0.2.0)",
+			req: &v1alpha1.MountRequest{
+				Attributes: `{"sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+			},
+			getAnnotationsFunc: func(namespace string, podName string) (map[string]string, error) {
+				return map[string]string{
+					"conjur.org/secrets":    "invalid",
+					"some-other-annotation": "some-value",
+				}, nil
 			},
 			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
 				assert.Nil(t, resp)
@@ -101,8 +135,13 @@ func TestMount(t *testing.T) {
 		{
 			description: "throws error decoding invalid file permissions",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"secrets":"- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Attributes: `{"sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
 				Permission: "abc",
+			},
+			getAnnotationsFunc: func(namespace string, podName string) (map[string]string, error) {
+				return map[string]string{
+					"conjur.org/secrets": "- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n",
+				}, nil
 			},
 			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
 				assert.Nil(t, resp)
@@ -112,9 +151,14 @@ func TestMount(t *testing.T) {
 		{
 			description: "throws error when conjur client fails",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"secrets":"- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Attributes: `{"sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
 				Permission: "777",
 				TargetPath: "/some/path",
+			},
+			getAnnotationsFunc: func(namespace string, podName string) (map[string]string, error) {
+				return map[string]string{
+					"conjur.org/secrets": "- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n",
+				}, nil
 			},
 			conjurFactory: func(baseURL, authnID, account, identity, sslCert string) conjur.Client {
 				return &mockConjurClient{
@@ -128,9 +172,9 @@ func TestMount(t *testing.T) {
 			},
 		},
 		{
-			description: "happy path",
+			description: "happy path (v0.1.0)",
 			req: &v1alpha1.MountRequest{
-				Attributes: `{"secrets":"- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Attributes: `{"conjur.org/configurationVersion":"0.1.0","secrets":"- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n","sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
 				Permission: "777",
 				TargetPath: "/some/path",
 			},
@@ -169,11 +213,59 @@ func TestMount(t *testing.T) {
 				})
 			},
 		},
+		{
+			description: "happy path (v0.2.0)",
+			req: &v1alpha1.MountRequest{
+				Attributes: `{"sslCertificate":"certificate content","account":"default","applianceUrl":"https://my.conjur.com","authnId":"authn-jwt/instance","identity":"botApp","csi.storage.k8s.io/serviceAccount.tokens":"{\"conjur\":{\"token\":\"sometoken\",\"expirationTimestamp\":\"2123-01-01T01:01:01Z\"}}"}`,
+				Permission: "777",
+				TargetPath: "/some/path",
+			},
+			conjurFactory: func(baseURL, authnID, account, identity, sslCert string) conjur.Client {
+				return &mockConjurClient{
+					resp: map[string][]byte{
+						"conjur/path/A": []byte("contentA"),
+						"conjur/path/B": []byte("contentB"),
+					},
+					err: nil,
+				}
+			},
+			getAnnotationsFunc: func(namespace string, podName string) (map[string]string, error) {
+				return map[string]string{
+					"conjur.org/secrets":    "- \"file/path/A\": \"conjur/path/A\"\n- \"file/path/B\": \"conjur/path/B\"\n",
+					"some-other-annotation": "some-value",
+				}, nil
+			},
+			assertions: func(t *testing.T, resp *v1alpha1.MountResponse, err error) {
+				assert.Nil(t, err)
+
+				assert.Len(t, resp.ObjectVersion, 2)
+				assert.Len(t, resp.Files, 2)
+
+				assert.Contains(t, resp.ObjectVersion, &v1alpha1.ObjectVersion{
+					Id:      "conjur/path/A",
+					Version: "1",
+				})
+				assert.Contains(t, resp.ObjectVersion, &v1alpha1.ObjectVersion{
+					Id:      "conjur/path/B",
+					Version: "1",
+				})
+				assert.Contains(t, resp.Files, &v1alpha1.File{
+					Path:     "file/path/A",
+					Mode:     int32(777),
+					Contents: []byte("contentA"),
+				})
+				assert.Contains(t, resp.Files, &v1alpha1.File{
+					Path:     "file/path/B",
+					Mode:     int32(777),
+					Contents: []byte("contentB"),
+				})
+			},
+		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
-			resp, err := mountWithDeps(context.TODO(), tc.req, tc.conjurFactory)
+			resp, err := mountWithDeps(context.TODO(), tc.req, tc.conjurFactory, tc.getAnnotationsFunc)
 			tc.assertions(t, resp, err)
 		})
 	}